Add files via upload
The Art store Aims to provide Online platform for art Collection to explore, discover and 
artwork from around the world. The website is designed to offer an experience showcasing a 
diverse range of artworks across various mediums and styles. 
our website's foundation is built upon HTML5 and CSS where HTML5 forms the structural 
backbone, defining and organizing the meaning of the content that brings these masterpieces 
to the website.
while CSS creates the structure with captivating visuals, making the exploration of artworks a 
delightful theme. 
From classic paintings to modern digital creations, Our Art Store provide a gallery to witness 
the harmonious blend of artistry with the technology includes HTML and CSS into our 
Website which builds architects, framing and enhancing the Quality. inviting you to discover, 
appreciate, and revel in the beauty of art through a digital lens of our creation.